Remove string-only references to location_* generated columns

+++ platform/lead.go
@@ -48,7 +48,7 @@ func leadCreate(ctx context.Context, txn db.Ex, user User, signal_id int32, site
 	}
 	return lead, nil
 }
-func leadsBySiteID(ctx context.Context, site_ids []int32) (map[int32][]*types.Lead, error) {
+func leadsBySiteID(ctx context.Context, site_ids []int64) (map[int32][]*types.Lead, error) {
 	rows, err := bob.All(ctx, db.PGInstance.BobDB, psql.Select(
 		sm.Columns(
 			models.Leads.Columns.ID.As("id"),
@@ -77,7 +77,7 @@ func leadsBySiteID(ctx context.Context, site_ids []int32) (map[int32][]*types.Le
 	}
 	results := make(map[int32][]*types.Lead, len(site_ids))
 	for _, site_id := range site_ids {
-		results[site_id] = make([]*types.Lead, 0)
+		results[int32(site_id)] = make([]*types.Lead, 0)
 	}
 	for _, row := range rows {
 		leads, ok := results[row.SiteID]
